unfify model loading
